# documentation/src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080/api
    description: Generated server url
...
paths:
...
  /v1/questionnaireresponse:
    get:
      tags:
        - QuestionnaireResponse
      operationId: getQuestionnaireResponsesByStatus
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              enum:
                - NOT_EXAMINED
                - UNDER_EXAMINATION
                - EXAMINED
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuestionnaireResponseDto'
    post:
      tags:
        - QuestionnaireResponse
      operationId: createQuestionnaireResponse
      parameters:
        - name: questionnaireResponse
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/QuestionnaireResponseDto'
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
  /v1/patient:
    get:
      tags:
        - Patient
      operationId: getPatient
      parameters:
        - name: cpr
          in: query
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PatientDto'
    post:
      tags:
        - Patient
      operationId: createPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
  /v1/questionnaireresponse/{id}:
    patch:
      tags:
        - QuestionnaireResponse
      operationId: patchQuestionnaireResponse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartialUpdateQuestionnaireResponseRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
...
  /v1/questionnaireresponse/{carePlanId}:
    get:
      tags:
        - QuestionnaireResponse
      operationId: getQuestionnaireResponsesByCarePlanId
      parameters:
        - name: carePlanId
          in: path
          required: true
          schema:
            type: string
        - name: questionnaireIds
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginatedListQuestionnaireResponseDto'
...
components:
  schemas:
    ErrorDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        errorCode:
          type: integer
          format: int32
        errorText:
          type: string
    EnableWhen:
      type: object
      properties:
        answer:
          $ref: '#/components/schemas/AnswerDto'
        operator:
          type: string
          enum:
            - EQUAL
            - GREATER_THAN
            - LESS_THAN
            - GREATER_OR_EQUAL
            - LESS_OR_EQUAL
    MeasurementTypeDto:
      type: object
      properties:
        system:
          type: string
        code:
          type: string
        display:
          type: string
    Option:
      type: object
      properties:
        option:
          type: string
        comment:
          type: string
    QuestionDto:
      type: object
      properties:
        linkId:
          type: string
        text:
          type: string
        abbreviation:
          type: string
        helperText:
          type: string
        required:
          type: boolean
        questionType:
          type: string
          enum:
            - CHOICE
            - INTEGER
            - QUANTITY
            - STRING
            - BOOLEAN
            - DISPLAY
            - GROUP
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
        enableWhen:
          type: array
          items:
            $ref: '#/components/schemas/EnableWhen'
        thresholds:
          type: array
          items:
            $ref: '#/components/schemas/ThresholdDto'
        measurementType:
          $ref: '#/components/schemas/MeasurementTypeDto'
        subQuestions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionDto'
        parentQuestion:
          $ref: '#/components/schemas/QuestionDto'
        deprecated:
          type: boolean
...
    ThresholdDto:
      type: object
      properties:
        questionId:
          type: string
        type:
          type: string
          enum:
            - NORMAL
            - ABNORMAL
            - CRITICAL
        valueBoolean:
          type: boolean
        valueQuantityLow:
          type: number
          format: double
        valueQuantityHigh:
          type: number
          format: double
        valueOption:
          type: string
...
    ContactDetailsDto:
      type: object
      properties:
        street:
          type: string
        postalCode:
          type: string
        country:
          type: string
        city:
          type: string
        primaryPhone:
          type: string
        secondaryPhone:
          type: string
    PatientDto:
      type: object
      properties:
        givenName:
          type: string
        familyName:
          type: string
        cpr:
          type: string
        customUserName:
          type: string
        patientContactDetails:
          $ref: '#/components/schemas/ContactDetailsDto'
        primaryRelativeName:
          type: string
        primaryRelativeAffiliation:
          type: string
        primaryRelativeContactDetails:
          $ref: '#/components/schemas/ContactDetailsDto'
        additionalRelativeContactDetails:
          type: array
          items:
            $ref: '#/components/schemas/ContactDetailsDto'
    AnswerDto:
      type: object
      properties:
        linkId:
          type: string
        value:
          type: string
        answerType:
          type: string
          enum:
            - INTEGER
            - STRING
            - BOOLEAN
            - QUANTITY
            - GROUP
        subAnswers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerDto'
    QuestionAnswerPairDto:
      type: object
      properties:
        question:
          $ref: '#/components/schemas/QuestionDto'
        answer:
          $ref: '#/components/schemas/AnswerDto'
    QuestionnaireResponseDto:
      required:
        - id
      type: object
      properties:
        id:
          type: string
          description: Id of the resource
          example: CarePlan/10
        questionnaireId:
          type: string
        carePlanId:
          type: string
        questionnaireName:
          type: string
        questionAnswerPairs:
          type: array
          items:
            $ref: '#/components/schemas/QuestionAnswerPairDto'
        answered:
          type: string
          format: date-time
        examinationStatus:
          type: string
          enum:
            - NOT_EXAMINED
            - UNDER_EXAMINATION
            - EXAMINED
        triagingCategory:
          type: string
          enum:
            - GREEN
            - YELLOW
            - RED
        patient:
          $ref: '#/components/schemas/PatientDto'
        planDefinitionTitle:
          type: string
...
    CreatePatientRequest:
      type: object
      properties:
        patient:
          $ref: '#/components/schemas/PatientDto'
...
    PartialUpdateQuestionnaireResponseRequest:
      type: object
      properties:
        examinationStatus:
          type: string
          enum:
            - NOT_EXAMINED
            - UNDER_EXAMINATION
            - EXAMINED
...
    PaginatedListQuestionnaireResponseDto:
      type: object
      properties:
        total:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
        list:
          type: array
          items:
            $ref: '#/components/schemas/QuestionnaireResponseDto'
...